Ling headspider needs an open turf to transform.

+++ code/modules/antagonists/changeling/powers/headcrab.dm
@@ -19,6 +19,10 @@
 			if(other_ling?.isabsorbing)
 				to_chat(user, "<span class='warning'>Our last resort is being disrupted by another changeling!</span>")
 				return
+	var/turf/T = user.loc
+	if(!isopenturf(T))
+		to_chat(user, "<span class='warning'>You need an open space to become a headslug!</span>")
+		return
 	if(alert("Are we sure we wish to kill ourself and create a headslug?",,"Yes", "No") == "No")
 		return
 	..()
@@ -47,7 +51,7 @@
 		cp.Remove(user)
 	user.gib()
 	. = TRUE
-	var/mob/living/simple_animal/hostile/headcrab/crab = new(turf)
+	var/mob/living/simple_animal/hostile/headcrab/crab = new(T)
 	for(var/obj/item/organ/I in organs)
 		I.forceMove(crab)
 	crab.origin = M
